namada_shielded_token
Module masp

MASP verification wrappers.

Re-exports

pub use shielded_wallet::ShieldedWallet;

pub use crate::masp::shielded_sync::ShieldedSyncConfigBuilder;

pub use crate::validation::partial_deauthorize;

pub use crate::validation::preload_verifying_keys;

pub use crate::validation::PVKs;

pub use crate::validation::CONVERT_NAME;

pub use crate::validation::ENV_VAR_MASP_PARAMS_DIR;

pub use crate::validation::OUTPUT_NAME;

pub use crate::validation::SPEND_NAME;

Modules

fs

Implementation of MASP functionality depending on a standard filesystem

shielded_wallet

The shielded wallet implementation

utils

Helper functions and types

Structs

Dispatcher

Shielded sync message dispatcher.

DispatcherCache

Shielded sync cache.

MaspChange

a masp change

MaspDataLog

Data to log the error of a batch of masp transactions

MaspDataLogEntry

Data to log the error of a single masp transaction

MaspFeeData

The data for a masp fee payment

MaspLocalTaskEnv

A task env whose backing thread-pool uses a no-op panic handler. Custom for MASP dispatchers.

MaspSourceTransferData

The data for a masp transfer relative to a given source

MaspTargetTransferData

The data for a masp transfer relative to a given target

MaspTokenRewardData

Shielded pool data for a token

MaspTransferData

The data for a single masp transfer

MaspTxReorderedData

ShieldedSyncConfig

A configuration used to tune the concurrency parameters of the shielded sync and the client used to fetch data.

ShieldedTransfer

Shielded transfer

WalletMap

Freeze a Builder into the format necessary for inclusion in a Tx. This is the format used by hardware wallets to validate a MASP Transaction.

Enums

ContextSyncStatus

The possible sync states of the shielded context

TransferErr

A return type for gen_shielded_transfer

Constants

ENV_VAR_MASP_TEST_SEED

Randomness seed for MASP integration tests to build proofs with deterministic rng.

NETWORK

The network to use for MASP

Traits

ShieldedUtils

Abstracts platform specific details away from the logic of shielded pool operations.

Functions

find_valid_diversifier

Generate a valid diversifier, i.e. one that has a diversified base. Return also this diversified base.

to_viewing_key

Make a ViewingKey that can view notes encrypted by given ExtendedSpendingKey

Type Aliases

Conversions

Represents the amount used of different conversions

MaspAmount

a masp amount

NoteIndex

Maps a shielded tx to the index of its first output note.

SpentNotesTracker

A type tracking the notes used to construct a shielded transfer. Used to avoid reusing the same notes multiple times which would lead to an invalid transaction

TransactionDelta

Represents the changes that were made to a list of shielded accounts

TransferDelta

Represents the changes that were made to a list of transparent accounts

WitnessMap

Maps the note index (in the commitment tree) to a witness