namada_shielded_token::masp::shielded_wallet
Trait ShieldedApi

pub trait ShieldedApi<U: ShieldedUtils + MaybeSend + MaybeSync>: ShieldedQueries<U> {

    // Provided methods
    async fn precompute_asset_types<C: Client + Sync>(
        &mut self,
        client: &C,
        tokens: Vec<&Address>,
    ) -> Result<(), Error> { ... }
    async fn decode_asset_type<C: Client + Sync>(
        &mut self,
        client: &C,
        asset_type: AssetType,
    ) -> Option<AssetData> { ... }
    async fn query_allowed_conversion<'a, C: Client + Sync>(
        &'a mut self,
        client: &C,
        asset_type: AssetType,
        conversions: &'a mut Conversions,
    ) -> Result<(), Error> { ... }
    async fn compute_exchanged_amount(
        &mut self,
        client: &(impl Client + Sync),
        io: &impl Io,
        input: I128Sum,
        conversions: Conversions,
    ) -> Result<(I128Sum, Conversions), Error> { ... }
    async fn compute_exchanged_balance(
        &mut self,
        client: &(impl Client + Sync),
        io: &impl Io,
        vk: &ViewingKey,
    ) -> Result<Option<I128Sum>, Error> { ... }
    async fn is_amount_required(
        &mut self,
        client: &(impl Client + Sync),
        src: ValueSum<(MaspDigitPos, Address), i128>,
        dest: ValueSum<(MaspDigitPos, Address), i128>,
        delta: I128Sum,
    ) -> Option<ValueSum<(MaspDigitPos, Address), i128>> { ... }
    async fn estimate_next_epoch_rewards(
        &mut self,
        context: &impl NamadaIo,
        raw_balance: &I128Sum,
    ) -> Result<DenominatedAmount, Error> { ... }
    async fn collect_unspent_notes(
        &mut self,
        context: &impl NamadaIo,
        spent_notes: &mut SpentNotesTracker,
        sk: PseudoExtendedKey,
        target: ValueSum<(MaspDigitPos, Address), i128>,
    ) -> Result<(I128Sum, Vec<(Diversifier, Note, MerklePath<Node>)>, Conversions), Error> { ... }
    async fn decode_combine_sum_to_epoch<C: Client + Sync>(
        &mut self,
        client: &C,
        amt: I128Sum,
        target_epoch: MaspEpoch,
    ) -> (ValueSum<Address, Change>, I128Sum) { ... }
    async fn decode_combine_sum<C: Client + Sync>(
        &mut self,
        client: &C,
        amt: I128Sum,
    ) -> (MaspAmount, I128Sum) { ... }
    async fn decode_sum<C: Client + Sync>(
        &mut self,
        client: &C,
        amt: I128Sum,
    ) -> (ValueSum<(AssetType, AssetData), i128>, I128Sum) { ... }
    async fn gen_shielded_transfer(
        &mut self,
        context: &impl NamadaIo,
        data: Vec<MaspTransferData>,
        fee_data: Option<MaspFeeData>,
        expiration: Option<DateTimeUtc>,
        bparams: &mut impl BuildParams,
    ) -> Result<Option<ShieldedTransfer>, TransferErr> { ... }
    async fn get_denom(
        client: &(impl Client + Sync),
        denoms: &mut HashMap<Address, Denomination>,
        token: &Address,
    ) -> Result<Denomination, TransferErr> { ... }
    async fn reorder_data_for_masp_transfer(
        context: &impl NamadaIo,
        data: Vec<MaspTransferData>,
        fee_data: Option<MaspFeeData>,
    ) -> Result<Option<MaspTxReorderedData>, TransferErr> { ... }
    async fn compute_change(
        &mut self,
        client: &(impl Client + Sync),
        added_amt: I128Sum,
        required_amt: ValueSum<(MaspDigitPos, Address), i128>,
        denoms: &mut HashMap<Address, Denomination>,
    ) -> Result<I128Sum, TransferErr> { ... }
    async fn add_inputs(
        &mut self,
        context: &impl NamadaIo,
        builder: &mut Builder<Network, PseudoExtendedKey>,
        source: &TransferSource,
        token: &Address,
        amount: &Amount,
        epoch: MaspEpoch,
        denoms: &mut HashMap<Address, Denomination>,
        notes_tracker: &mut SpentNotesTracker,
    ) -> Result<Option<I128Sum>, TransferErr> { ... }
    async fn add_outputs(
        &mut self,
        context: &impl NamadaIo,
        builder: &mut Builder<Network, PseudoExtendedKey>,
        source: Option<TransferSource>,
        target: &TransferTarget,
        token: Address,
        amount: Amount,
        epoch: MaspEpoch,
        denoms: &mut HashMap<Address, Denomination>,
    ) -> Result<(), TransferErr> { ... }
    async fn get_asset_type<C: Client + Sync>(
        &mut self,
        client: &C,
        decoded: &mut AssetData,
    ) -> Result<AssetType, Error> { ... }
    async fn convert_namada_amount_to_masp<C: Client + Sync>(
        &mut self,
        client: &C,
        epoch: MaspEpoch,
        token: &Address,
        denom: Denomination,
        val: Amount,
    ) -> Result<([AssetType; 4], U64Sum), Error> { ... }

}

The methods of the shielded wallet that depend on the ShieldedQueries trait. These cannot be overridden downstream.

Provided Methods

async fn precompute_asset_types<C: Client + Sync>( &mut self, client: &C, tokens: Vec<&Address>, ) -> Result<(), Error>

Use the addresses already stored in the wallet to precompute as many asset types as possible.

async fn decode_asset_type<C: Client + Sync>( &mut self, client: &C, asset_type: AssetType, ) -> Option<AssetData>

Query the ledger for the decoding of the given asset type and cache it if it is found.

async fn query_allowed_conversion<'a, C: Client + Sync>( &'a mut self, client: &C, asset_type: AssetType, conversions: &'a mut Conversions, ) -> Result<(), Error>

Query the ledger for the conversion that is allowed for the given asset type and cache it. The target epoch must be greater than or equal to the asset’s epoch. If the target epoch is None, then convert to the latest epoch.

async fn compute_exchanged_amount( &mut self, client: &(impl Client + Sync), io: &impl Io, input: I128Sum, conversions: Conversions, ) -> Result<(I128Sum, Conversions), Error>

Convert the given amount into the latest asset types whilst making a note of the conversions that were used. Note that this function does not assume that allowed conversions from the ledger are expressed in terms of the latest asset types.

async fn compute_exchanged_balance( &mut self, client: &(impl Client + Sync), io: &impl Io, vk: &ViewingKey, ) -> Result<Option<I128Sum>, Error>

Compute the total unspent notes associated with the viewing key in the context and express that value in terms of the currently timestamped asset types. If the key is not in the context, then we do not know the balance and hence we return None.

async fn is_amount_required( &mut self, client: &(impl Client + Sync), src: ValueSum<(MaspDigitPos, Address), i128>, dest: ValueSum<(MaspDigitPos, Address), i128>, delta: I128Sum, ) -> Option<ValueSum<(MaspDigitPos, Address), i128>>

Determine if using the current note would actually bring us closer to our target. Returns the contribution of the current note to the target if so.

async fn estimate_next_epoch_rewards( &mut self, context: &impl NamadaIo, raw_balance: &I128Sum, ) -> Result<DenominatedAmount, Error>

We estimate the next epoch rewards accumulated by the assets included in the provided raw balance, i.e. a balance based only on the available notes, without accounting for any conversions. The estimation is done by assuming the same rewards rate on each asset as in the latest masp epoch.

async fn collect_unspent_notes( &mut self, context: &impl NamadaIo, spent_notes: &mut SpentNotesTracker, sk: PseudoExtendedKey, target: ValueSum<(MaspDigitPos, Address), i128>, ) -> Result<(I128Sum, Vec<(Diversifier, Note, MerklePath<Node>)>, Conversions), Error>

Collect enough unspent notes in this context to exceed the given amount of the specified asset type. Return the total value accumulated plus notes and the corresponding diversifiers/merkle paths that were used to achieve the total value. Updates the changes map.

async fn decode_combine_sum_to_epoch<C: Client + Sync>( &mut self, client: &C, amt: I128Sum, target_epoch: MaspEpoch, ) -> (ValueSum<Address, Change>, I128Sum)

Convert an amount whose units are AssetTypes to one whose units are Addresses that they decode to. All asset types not corresponding to the given epoch are ignored.

async fn decode_combine_sum<C: Client + Sync>( &mut self, client: &C, amt: I128Sum, ) -> (MaspAmount, I128Sum)

Convert an amount whose units are AssetTypes to one whose units are Addresses that they decode to and combine the denominations.

async fn decode_sum<C: Client + Sync>( &mut self, client: &C, amt: I128Sum, ) -> (ValueSum<(AssetType, AssetData), i128>, I128Sum)

Convert an amount whose units are AssetTypes to one whose units are Addresses that they decode to.

async fn gen_shielded_transfer( &mut self, context: &impl NamadaIo, data: Vec<MaspTransferData>, fee_data: Option<MaspFeeData>, expiration: Option<DateTimeUtc>, bparams: &mut impl BuildParams, ) -> Result<Option<ShieldedTransfer>, TransferErr>

Make shielded components to embed within a Transfer object. If no shielded payment address nor spending key is specified, then no shielded components are produced. Otherwise, a transaction containing nullifiers and/or note commitments are produced. Dummy transparent UTXOs are sometimes used to make transactions balanced, but it is understood that transparent account changes are effected only by the amounts and signatures specified by the containing Transfer object.

async fn get_denom( client: &(impl Client + Sync), denoms: &mut HashMap<Address, Denomination>, token: &Address, ) -> Result<Denomination, TransferErr>

Either get the denomination from the cache or query it

async fn reorder_data_for_masp_transfer( context: &impl NamadaIo, data: Vec<MaspTransferData>, fee_data: Option<MaspFeeData>, ) -> Result<Option<MaspTxReorderedData>, TransferErr>

Group all the information for every source/token and target/token couple, and extract the denominations for all the tokens involved (expect the one involved in the fees if needed). This step is required so that we can collect the amount required for every couple and pass it to the appropriate function so that notes can be collected based on the correct amount.

async fn compute_change( &mut self, client: &(impl Client + Sync), added_amt: I128Sum, required_amt: ValueSum<(MaspDigitPos, Address), i128>, denoms: &mut HashMap<Address, Denomination>, ) -> Result<I128Sum, TransferErr>

Computes added_amt - required_amt taking care of denominations and asset decodings. Error out if required_amt is not less than added_amt.

async fn add_inputs( &mut self, context: &impl NamadaIo, builder: &mut Builder<Network, PseudoExtendedKey>, source: &TransferSource, token: &Address, amount: &Amount, epoch: MaspEpoch, denoms: &mut HashMap<Address, Denomination>, notes_tracker: &mut SpentNotesTracker, ) -> Result<Option<I128Sum>, TransferErr>

Add the necessary transaction inputs to the builder. The supplied epoch must be the current epoch.

async fn add_outputs( &mut self, context: &impl NamadaIo, builder: &mut Builder<Network, PseudoExtendedKey>, source: Option<TransferSource>, target: &TransferTarget, token: Address, amount: Amount, epoch: MaspEpoch, denoms: &mut HashMap<Address, Denomination>, ) -> Result<(), TransferErr>

Add the necessary transaction outputs to the builder

async fn get_asset_type<C: Client + Sync>( &mut self, client: &C, decoded: &mut AssetData, ) -> Result<AssetType, Error>

Get the asset type with the given epoch, token, and denomination. If it does not exist in the protocol, then remove the timestamp. Make sure to store the derived AssetType so that future decoding is possible.

async fn convert_namada_amount_to_masp<C: Client + Sync>( &mut self, client: &C, epoch: MaspEpoch, token: &Address, denom: Denomination, val: Amount, ) -> Result<([AssetType; 4], U64Sum), Error>

Convert Namada amount and token type to MASP equivalents

Object Safety

This trait is not object safe.