namada_shielded_token
Module storage_key

Shielded token storage keys

Constants

MASP_ASSETS_HASH_KEY

The key for the asset map

MASP_BASE_NATIVE_PRECISION_KEY

The key for the base native precision

MASP_CONVERSIONS_KEY

Key segment prefix for the conversions

MASP_CONVERT_ANCHOR_KEY

Key segment prefix for the convert anchor

MASP_KD_GAIN_KEY

The key for the nominal derivative gain of a shielded pool for a given asset

MASP_KP_GAIN_KEY

The key for the nominal proportional gain of a shielded pool for a given asset

MASP_LAST_INFLATION_KEY

Last calculated inflation value handed out

MASP_LAST_LOCKED_AMOUNT_KEY

The last locked amount

MASP_LOCKED_AMOUNT_TARGET_KEY

The key for the locked amount target for a given asset

MASP_MAX_REWARD_RATE_KEY

The key for the max reward rate for a given asset

MASP_NOTE_COMMITMENT_ANCHOR_PREFIX

Key segment prefix for the note commitment anchor

MASP_NOTE_COMMITMENT_TREE_KEY

Key segment prefix for the note commitment merkle tree

MASP_NULLIFIERS_KEY

Key segment prefix for the nullifiers

MASP_REWARD_PRECISION_KEY

The key for the reward precision for a given asset

MASP_SCHEDULED_BASE_NATIVE_PRECISION_KEY

Key segment prefix for the scheduled base native precisions

MASP_SCHEDULED_REWARD_PRECISION_KEY

Key segment prefix for the scheduled reward precisions

MASP_TOKEN_MAP_KEY

The key for the token map

MASP_TOTAL_REWARDS

The key for the total inflation rewards minted by MASP

MASP_UNDATED_BALANCE_KEY

The key for the masp reward balance

Functions

is_masp_balance_key

Check if the given storage key is MASP transparent balance key

is_masp_base_native_precision_key

Check if the given storage key is a masp base native precision key

is_masp_commitment_anchor_key

Check if the given key is a masp commitment anchor

is_masp_commitment_tree_key

Check if the given storage key is a masp commitment tree key

is_masp_conversion_key

Check if the given storage key is a masp conversion key

is_masp_governance_key

Check if the given storage key is allowed to be touched by a governance proposal

is_masp_key

Check if the given storage key is a masp key

is_masp_nullifier_key

Check if the given storage key is a masp nullifier key

is_masp_scheduled_base_native_precision_key

Check if the given storage key is a scheduled base native precision

is_masp_scheduled_reward_precision_key

Check if the given storage key is the scheduled reward precision of a token

is_masp_token_map_key

Check if the given storage key is a masp token map key

is_masp_transfer_key

Check if the given storage key is allowed to be touched by a masp transfer

is_masp_undated_balance_key

Check if the given storage key is the undated balance of a token

masp_assets_hash_key

Get the key for the masp assets’ hash

masp_base_native_precision_key

Get the key for the masp base native precision map

masp_commitment_anchor_key

Get a key for a masp commitment tree anchor

masp_commitment_tree_key

Get the key for the masp commitment tree

masp_conversion_key

Get a key for a masp conversion at the given epoch

masp_conversion_key_prefix

Get the key prefix for masp conversions at the given epoch

masp_convert_anchor_key

Get the key for the masp convert tree anchor

masp_kd_gain_key

Obtain the nominal derivative key for the given token

masp_kp_gain_key

Obtain the nominal proportional key for the given token

masp_last_inflation_key

Obtain the storage key for the last inflation of a token

masp_last_locked_amount_key

Obtain the storage key for the last locked amount of a token

masp_locked_amount_target_key

Obtain the locked target amount key for the given token

masp_max_reward_rate_key

The max reward rate key for the given token

masp_nullifier_key

Get a key for a masp nullifier

masp_reward_precision_key

The shielded reward precision key for the given token

masp_scheduled_base_native_precision_key

Get the storage key for scheduled masp base native precision at the given epoch

masp_scheduled_reward_precision_key

Obtain the storage key for the scheduled reward precision of a token

masp_scheduled_reward_precision_key_prefix

Get the key prefix for scheduled masp reward precision at the given epoch

masp_token_map_key

Get the key for the masp token map

masp_total_rewards

The max reward rate key for the given token

masp_undated_balance_key

Obtain the storage key for the undated balance of a token